sample e2e test structure
