+++ test/e2e/platform_persistence_resillence_test.go
@@ -1,0 +1,1 @@
+package e2e
