+++ test/e2e/platform_resillence_test.go
