+++ test/e2e/platform_tiered_storage_test.go
@@ -1,0 +1,1 @@
+package e2e
